Clean, get_file_info_files(), back to return() calls
Again I got errors "Object of type 'S4' is not
subsettable". Explicitly using "else {NULL}" made
these errors disappear. However, I decided to use
return() again within the reactive expressions. It
should be ok since the expression that is passed to
reactive() will become the body of a function and it
will be that function that is returned from.